+++ hcloud/certificate.go
@@ -51,6 +51,9 @@ func (c *CertificateClient) GetByID(ctx context.Context, id int) (*Certificate, 
 
 // GetByName retrieves a Certificate by its name. If the Certificate does not exist, nil is returned.
 func (c *CertificateClient) GetByName(ctx context.Context, name string) (*Certificate, *Response, error) {
+	if name == "" {
+		return nil, nil, nil
+	}
 	Certificate, response, err := c.List(ctx, CertificateListOpts{Name: name})
 	if len(Certificate) == 0 {
 		return nil, response, err