+++ hcloud/floating_ip.go
@@ -76,6 +76,9 @@ func (c *FloatingIPClient) GetByID(ctx context.Context, id int) (*FloatingIP, *R
 
 // GetByName retrieves a Floating IP by its name. If the Floating IP does not exist, nil is returned.
 func (c *FloatingIPClient) GetByName(ctx context.Context, name string) (*FloatingIP, *Response, error) {
+	if name == "" {
+		return nil, nil, nil
+	}
 	floatingIPs, response, err := c.List(ctx, FloatingIPListOpts{Name: name})
 	if len(floatingIPs) == 0 {
 		return nil, response, err