+++ hcloud/server_type.go
@@ -69,6 +69,9 @@ func (c *ServerTypeClient) GetByID(ctx context.Context, id int) (*ServerType, *R
 
 // GetByName retrieves a server type by its name. If the server type does not exist, nil is returned.
 func (c *ServerTypeClient) GetByName(ctx context.Context, name string) (*ServerType, *Response, error) {
+	if name == "" {
+		return nil, nil, nil
+	}
 	serverTypes, response, err := c.List(ctx, ServerTypeListOpts{Name: name})
 	if len(serverTypes) == 0 {
 		return nil, response, err